Add password authentication for noVNC
This adds the option to authenticate users using noVNC. The noVNC website has to be embedded as CoWebsite. Then, the boolean option 'authenticateNoVNC', which is added by this commit, has to be set to true. The password has to be saved in the environment file, named 'NOVNC_PASSWORD'. When loading the website, the password is sent to the user by pusher and appended to the noVNC URL.